1884-dev-trafficincidentagent: Add timer
