+++ Agents/TrafficIncidentAgent/TrafficIncidentAgentCode/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>uk.ac.cam.cares.jps</groupId>
     <artifactId>traffic-incident-agent</artifactId>
-    <version>1.4.0</version>
+    <version>1.5.0</version>
     <packaging>war</packaging>
     
     
@@ -130,21 +130,21 @@
             <version>4.5.10</version>
         </dependency>
 
-        <!-- Logging -->
+        <!-- Logging, versions pulled from parent -->
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-api</artifactId>
-            <version>2.17.0</version>
+            <version>2.23.0</version>
         </dependency>
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-core</artifactId>
-            <version>2.17.0</version>
+            <version>2.23.0</version>
         </dependency>
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-web</artifactId>
-            <version>2.17.0</version>
+            <version>2.23.0</version>
         </dependency>
 
         <!-- Testing -->
@@ -205,7 +205,13 @@
         <dependency>
             <groupId>com.cmclinnovations</groupId>
             <artifactId>stack-clients</artifactId>
-            <version>1.28.0</version>
+            <version>1.37.3</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>ch.qos.logback</groupId>
+                    <artifactId>logback-classic</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
 
     </dependencies>
