+++ Agents/TrafficIncidentAgent/TrafficIncidentAgentCode/src/main/java/uk/ac/cam/cares/jps/agent/trafficincident/TrafficIncident.java
@@ -1,6 +1,11 @@
 package uk.ac.cam.cares.jps.agent.trafficincident;
 
+import java.util.UUID;
+
 public class TrafficIncident {
+    // TODO: to be updated
+    public static final String IRI_PREFIX = "https://www.theworldavatar.com/kg/ontoroad/TrafficIncident_";
+
     public long startTime;
     public long endTime;
     public String incidentType;
@@ -9,8 +14,21 @@ public class TrafficIncident {
     // message field can be updated during different call, depending on specific scenario
     public String message;
     public boolean status; // true : incident is ongoing
+    public String iri; // IRI_PREFIX + random UUID
 
     public TrafficIncident(String incidentType, double latitude, double longitude, String message, long startTime, boolean status) {
+        this.startTime = startTime;
+        this.endTime = 0;
+        this.incidentType = incidentType;
+        this.latitude = latitude;
+        this.longitude = longitude;
+        this.message = message;
+        this.status = status;
+        this.iri = TrafficIncident.IRI_PREFIX + UUID.randomUUID().toString();
+    }
+
+    public TrafficIncident(String iri, String incidentType, double latitude, double longitude, String message, long startTime, boolean status) {
+        this.iri = iri;
         this.startTime = startTime;
         this.endTime = 0;
         this.incidentType = incidentType;
@@ -31,11 +49,11 @@ public class TrafficIncident {
     @Override
     public String toString() {
         if (this.endTime == 0) {
-            return String.format("%s %s at latitude %f, longitude %f starting from %d", (this.status ? "Ongoing" : "Ended"), this.incidentType, this.latitude, this.longitude, this.startTime);
+            return String.format("%s %s %s at latitude %f, longitude %f starting from %d", (this.status ? "Ongoing" : "Ended"), this.iri, this.incidentType, this.latitude, this.longitude, this.startTime);
         } else {
-            return String.format("%s %s at latitude %f, longitude %f starting from %d to %d", (this.status ? "Ongoing" : "Ended"), this.incidentType, this.latitude, this.longitude, this.startTime, this.endTime);
+            return String.format("%s %s %s at latitude %f, longitude %f starting from %d to %d", (this.status ? "Ongoing" : "Ended"), this.iri, this.incidentType, this.latitude, this.longitude, this.startTime, this.endTime);
         }
-        
+
     }
 
     @Override
@@ -47,9 +65,9 @@ public class TrafficIncident {
             // under the assumption that TrafficIncident with same type, location and start time must be the same (message may get updated during different call)
             TrafficIncident other = (TrafficIncident) obj;
             return this.incidentType.equals(other.incidentType)
-                && this.latitude == other.latitude
-                && this.longitude == other.longitude
-                && this.startTime == other.startTime;
+                    && this.latitude == other.latitude
+                    && this.longitude == other.longitude
+                    && this.startTime == other.startTime;
         }
         return false;
     }
